Publicize server (#283)
Many internals looking to start working with the OPC-UA extractor have asked about servers to test against. The internal test server has been the answer thus far, but it was a little hard to work with, since it required modifying the code to run different components.

This PR aims to make it accessible through a simple CLI, which just exposes various debug functionality added over the years. It also publicizes the server during releases, so that it can be used without installing the .NET SDK.