[CDF-14271] Improve valuerank log (#288)
ValueRank is a value indicating the dimensionality of a variable. It is a number (-3, -2, -1, 0, 1, ++), where -3 to 0 have special meanings, and higher values is just the number of dimensions. Importantly we log these when timeseries are filtered out due to configuration, but just the numbers are a little unclear, so we should log their names proper.