[CDF-14285] Timeseries as events (#290)
System to generate events based on variable changes instead of datapoints.

This is useful in cases where complex event information is stored in a datapoint instead of as OPC-UA events. Sometimes a variable maps to an event more than a timeseries, i.e. if the variable actually is some alarm value or the like. This could be handled by triggers in CDF, but we might get more metadata from the value by doing it this way.