[CDF-14331] Browse callback (#295)
System for calling a CDF function after each browse/rebrowse. The parameters to the function are the number of assets/timeseries/relationships created and modified, and may refer to raw rows instead, depending. It also includes some data about the extraction to make it possible for a single function to be reused for multiple extractors (in theory, at least, the information can't hurt).

This required some slightly unpleasant refactoring of the pushers. The issue is that this should be localized to the CDFPusher, but pushing of relationships and assets/timeseries were completely separate. This wasn't really that useful to begin with, but we want to report success for these operations separately. This PR combines pushing of references and nodes, but returns success for the three operations separately.

For the MQTTPusher (which had to be changed since they have the same interface) this actually creates a small performance improvement.