Re-add restarting history on reconnect (#296)
This disappeared at some point, I think. In general we want to restart history when the extractor reconnects to the server, even if restart-on-reconnect is false.

This also fixes makes historyReader.Terminate do what it says (terminate any running history processes). Weird things might happen if this happens several times in sequence.

There is also the potential issue of what to do if history read fails entirely. This is kinda tricky to work with. WIth any other extractor we would crash immediately, but apparently some OPC-UA servers are so slow this is problematic.