Let max be history end time if set
This resolves the servi issue. Their problem was a noncompliant server
that returned no results if end-time was set to DateTime.MinValue, in
violation of the OPC-UA standard.

Fortunately, a simple workaround is to fix this (which is technically a
bug), then let them set end-time and start-time on their extractor to
some reasonable numbers.