[CDF-14843] Use test utils (#299)
Using the "MakeConsoleWork" class is unreliable and bad practice. With the newest version of the utils it is possible to use a logger with replaceable TestOutputHelper, so we use that instead.

This also updated the OPC-UA SDK, which broke a few tests since they now (correctly) set the initial state of values on the server to BadWaitingForInitialData.