Refactor config tool (#303)
Big PR, but it is mostly just splitting the code for UAServerExplorer into different files. The overall functionality is still the same, but I added a few extra bits of information to the summary.

The configuration tool had become difficult to maintain, since the UAServerExplorer file was so big. Really big files discourage additions. This helps by splitting the code up by function, although it is still all in one class, since we use internal UAClient methods.

I also moved logging of summary into the Summary class itself, which is just cleaner and more flexible.