[CDF-15208] Update to utils 1.0.0 - cron support (#307)
* Update packages

* Use utils log-to-string method

* Add cron expression support

* Reduce test verbosity a little

New test logging system plays better with xunit, so we no longer need
the -v, which now just creates an enormous amount of logs in jenkins.

* Fix tests

There were two issues here, caused by changes to the SDK

First: We test the number of results from reading types from OPC-UA,
since that is consistent. When the SDK adds more nodes, that number
needs to be updated, this is expected.

Seconds: They have apparently changed browse to be "fair", so results
are distributed over nodes, probably some parallelism internally or
something along those lines. It's a sensible change, but it conflicts
with the way we simulate the issue, and actually exposes a flaw in the
config tool. This helps alleviate that, though it does not solve it
fully.