Add github action for PRs (#308)
Separate pipelines for release and build. The release pipeline is pretty complicated, since we release so many different types of artifacts here.

Missing docker build for now, we need to decide whether it should be published to dockerhub or our internal docker repo.

Also needed to remove console logging as much as possible, since for some reason that is written globally when running xunit tests in github actions.