Filter references (#314)
Refactor reference mapping in a few ways.

 1. Move filtering to a common method in BaseNodeSource.
 2. Add more context to the dictionary of mapped nodes in state, and use this to determine the vertex type.

Some additional context for 2:

Originally we stored the entire object for each node, but this had the side effect of using a lot of memory, so we moved to using a checksum.

This works well, until we need information about nodes not in memory. This typically happens when we map a subset of the node hierarchy during periodic updates. When looking for references we need to know whether the target node is a timeseries or asset, as that will matter when references are mapped to relationships.

This PR stores this additional context. The memory use should be small, since the size of each MappedNode is fixed (NodeIds are already stored, and the rest of the members are value types).

The linux-only reconnect test also finally failed, so I got rid of it. It isn't ideal, but loss of network connection is almost impossible to test in CI, and the test has always been really unstable. Now it broke completely, and I don't want to go looking for a replacement. Instead I just stop and restart the server, which isn't exactly the same, but is close enough.