Refactor connection logic
The idea is to lower restart times and also clean up the code a bit.

Instead of restarting from zero when the extractor fails to connect, we
can restart inside a new SessionManager class.

This PR also improves the ForceRestart option to not quit the extractor
entirely, which greatly improves restart time.

Some users have been hesitant to use the ForceRestart option because it
is so slow, this should help them considerably.