Fix retries (#335)
The issue was that when silenced exceptions were given to HandleServiceResult, they were treated as unknown exceptions. We would rather just return them back.

This also contains a bit of server code to make testing this easier, though it's not something I would like to include in tests.

It also adds `BadNoCommunication` as a default status code.