Pass delete request to pushers (#344)
This ended up being somewhat more complicated than expected. The PushNodes methods in UAExtractor were not really that well suited to get a list of nodes to delete as well, and required a small rewrite. The system for caching nodes that failed during the initial push also required a little work. The result is (IMO) a lot cleaner, and makes a lot more sense.

Beyond that we now have an ExecuteDeletes method on IPusher, which can be implemented for CDFPusher to handle deletes.