Handle skipping a variable that is parent of another node
We skip variables for a variety of reasons. If MapVariableChildren is
enabled, variables with children are turned into both an object and a
variable, i.e.

Parent
  MyVariable

is turned into

Parent
  MyVariable (object)
    MyVariable
    Children of MyVariable

If we skipped MyVariable in AllowTsMap, we would skip the entire node.
This isn't ideal, as we would like the result to be

Parent
  MyVariable (object)
    Children of MyVariable

So that children are mapped.

This solves that. The fix is very simple, just move the call to
AllowTsMap into AddVariableToLists, only adding the variable to
_variable_ lists if AllowTSMap returns true, but always adding it to
_object_ lists.