Remove top level visited nodes logic
This is an old feature intended to reduce the load on rebrowse. In the
very early days of the extractor, we did not do any form of updates at
all, and only discovered new nodes. It was nice, then, to ignore nodes
already when browsing, so that we don't read the values of these nodes.

In general though, users almost always want updates, and the cost is not
that large. Reading attributes is usually the cheapest part of the
rebrowse process.

This PR removes checking for visited nodes entirely, except within the
current browse (we need to do this to avoid cycles). If we encounter a
node we have already visited we now pass that along to the callback
instead. This was necessary in the FDM branch in order to register all
relationships, if a node appears twice in the hierarchy (which is
valid!).