Remove top level visited nodes logic (#377)
First of a series of PRs porting changes made in the FDM branch to master. Both to reduce the eventual complexity of the FDM change, and to actually get these fixes in without introducing too many bugs.

This is an old feature intended to reduce the load on rebrowse. In the very early days of the extractor, we did not do any form of updates at all, and only discovered new nodes. It was nice, then, to ignore nodes already when browsing, so that we don't read the values of these nodes.

In general though, users almost always want updates, and the cost is not that large. Reading attributes is usually the cheapest part of the rebrowse process.

This PR removes checking for visited nodes entirely, except within the current browse (we need to do this to avoid cycles). If we encounter a node we have already visited we now pass that along to the callback instead. This was necessary in the FDM branch in order to register all relationships, if a node appears twice in the hierarchy (which is valid!).

Note that this is not a breaking change. It does not change the behavior of the extractor, since we still validate check for changes when writing to CDF. It will increase rebrowse time on servers with the default configuration.