Refactor node types (#382)
Huge refactor of how OPC-UA nodes are represented in the extractor.

When the extractor was first written, it only supported Objects and Variables, and only data point subscriptions for variables. This meant that Variables were effectively subtypes of Objects, and so we wrote them in as `UANode` and `UAVariable`.

Once we needed to look at the type structure of the OPC-UA server, this was added as various TypeManagers, which had their own specialized types.

Now we support data types, object types, variable types, and reference types in addition to objects and variables. This means that there is a clear overlap between the nodes we support in general, and specifically in type managers.

In order to improve the ergonomics of this, and make it easier to add behavior later, this PR adds new types representing each of the different node classes, and replaces the type managers with a single `TypeManager` class with much more generic behavior. This has a lot of consequences in the code, and simply fixing the build took a lot of work.