Replace pusher debug with dry-run option
This is a bit more general, the pusher debug option was always a bit
artificial. This is intended as a proper way to prevent the extractor
from persisting anything, while still letting users see what the
extractor would do.

This should mostly not access destinations at all. The exception is if
CDF is configured as source, which is intentional.