Allow sharing server between extractors
This was originally planned for later, but this also happens to solve
the issue that we need to reserve the namespace-index 1.

This is painful, because it means we need to translate node ID namespace
indices to a shared set used by the server.

The change should not be too complicated. The information necessary to
translate is kept in a NodeIdContext class, then we just need to make
sure to use that everywhere we did `NodeId.ToString()`