Refactor node sources (#498)
Major refactor to node sources.

The problem has been that each node source effectively needs to replicate the whole node ingestion logic. The common parts were included in an abstract base class, but that really just made things harder to understand, as it required anyone working on a source to understand what the base class was doing at all times.

This more or less flips the dependency, making the base class (now NodeHierarchyBuilder, and not a base class, a wrapper) call methods on sources as needed.

This also abstracts out the reading of types from the type manager, making it use the NodeSetSource when that is configured to contain type information, which before worked through a side channel.

There is certainly a lot more to be done, but this is a start.