Centralize subscription management (#500)
* Centralize subscription management

This moves all logic for creating and maintaining subscriptions out of
UAClient and into a dedicated subscription manager. This tidies up a few
different features related to subscriptions that were interacting
dangerously.

This is a bit less convenient for testing, but it does tidy up UAClient
a bit, which is good, since that class is too big.

* Fix config tool tests

* Fix hanging launcher tests

* Use flattened exception

* More review comments