Add cache for subscription states (#504)
This adds a cache containing the state of each subscription, and when it was last updated. It is effectively a very overkill solution for making subscription recreation wait a bit before they run, to help prevent subscriptions from being queued for deletion, then recover after being fully created.

The solution is a bit much, but it's relatively tidy, and it should work. This is pretty much impossible to test.