Separate session context from session (#508)
* Separate session context from session

This is a bit tricky, but it might allow the extractor to keep working
even when connection to the server is lost. It should generally be fine
to continue going with the _previous_ session context. If the setup is
such that this won't work, the extractor should be configured to restart
hard on connection lost. Most servers will work fine with this.

* Fix tests

* Make session context non-nullable

* Fix tests

* PR fixes

* Bump version