Refactor CDF writers to include deletes (#514)
This was meant to just be a PR to add deletes to FDM, but it became clear that the structure of the new writer setup was poorly suited for this, and there are also a few bugs in there. This refactors it (removing quite a few lines of code), and also moves deletes into the writers, which generally makes sense. The old approach didn't really make CDFPusher any more readable, which was the entire point.

Actually implementing FDM deletes will come in a separate PR.