Cap the amount of data buffered in the queues.
This should prevent issues if the network to CDF is down while the
extractor is reading history, somehow.

Technically this could cause issues if the queue is being flooded by
historical data to such a degree that live data is blocked and lost, but
in general we should never even encounter this issue.