Cap the amount of data buffered in the queues. (#637)
This should prevent issues if the network to CDF is down while the extractor is reading history, somehow.

Technically this could cause issues if the queue is being flooded by historical data to such a degree that live data is blocked and lost, but in general we should never even encounter this case at all, so I think that's a fair tradeoff. If the network is that flaky and wrong you can't expect everything to go well. In the end we still track ranges, so we shouldn't lose data.