Refactor SessionManager (#747)
The current solution is just too fragile and unstable. I wanted to avoid doing this, but I didn't see a way forward with the current implementation. This fundamentally changes a few things:

 - Abstract over reconnect/connect methods by creating an `IConnectionSource` interface.
 - Run a single loop to actually invoke `Connect`. This loop is notified from other threads through a condition variable and holds a lock over almost its entire body, which should help simplify the possible traces of the session manager, and make future bugs easier to fix and less likely to occur in the first place.

I have tested this as much as I am able, and of course there are some automated tests that still pass, but the issues we've had with the old implementation are really difficult to reproduce, since they relied heavily on highly specific timing.