Remove read extracted ranges (#765)
Removes the read-extracted-ranges feature completely.

There may be a little room for more refactoring in the UAExtractor class, but not a lot and that can wait for later. This makes state-storage required, sets the state-storage interval by default (you still need to set a location and type), and removes the read-extracted-ranges option everywhere.