Make update non-optional
When we added support for "updating" metadata in CDF, it was done in a very flexible way.

Now that we write to DM more, it really makes a lot less sense, since depending on where you write there is no way to _not_ update. We had some absolutely insane code to enable "updates" to Raw that can also be removed (no other extractors do anything like that).

In general, making updates non-optional greatly simplifies the extractor, and is almost certainly what pretty much every user wants.

Also identifies and fixes a bug related to the `Source` of arrays, this was never caught due to a hole in a test.

Fix update test