Remove the ability to have multiple or zero pushers (#768)
No improvement in LoC hurts me, but this is still an improvement in how easy the code is to understand. It just caused a lot of ripple effects.

The biggest change apart from the title is that I added separate methods to check if we can push to events/timeseries, and changed the return type for the push methods to be an enum.

This in general better matches how the extractor actually behaves. It also specifically tracks errors that we are confident we won't recover from without human intervention, which will be quite handy once we start implementing new extraction pipelines.