Patch for memory leak on event buffer
This is the same issue as the other memory leak, except caused by the
buffer for events instead of the one for datapoints...

The patch is even a little hackier, but I also realized that this
feature only actually does _anything_ if influxdb is enabled (since that
is the only target where we use the state of the target as state for
events).

Keep in mind that this entire feature is being removed completely in v3.
To avoid breaking existing deployments I'm keeping it, ref discussion
when we did this for the data point buffer, but I'm disabling it for
deployments without influx, since it never worked for those.