Always enable data-type-metadata (#786)
This feature adds a metadata field dataType to timeseries.

This is something users are extremely likely to want, and the theoretical cost is tiny (if any). It makes sense to enable by default.

For risk review: We are making a batch of breaking changes to the OPC-UA extractor on a feature branch. These are intended to simplify the extractor, in order to make it easier to use and more maintainable. This is the latest in a series of PRs that enable some config options by default.