Add options to adjust MaxNotificationsPerPublish
This is a server option, in practice, so all we do is send this number
to the server. Because the size of datapoint and event notifications can
be so different, I added a separate number for the two subscriptions.

We've rarely actually seen an issue with this, servers are free to send
however many notifications they want.