Set MaxMonitoredItemsPerCall on the client
We try to do our own batching here, but the client doesn't expose an
interface to allow us to do that. Apparently the current way we do it
can fail, or create way too many monitored items to be created at a
time. This should hopefully protect against this case.