Refactor string converter
This type is the result of years of feature creep, and it is incredibly
messy. This splits the ConvertToString method into ConvertToJson and
ConvertToString, which hopefully will make it a lot more readable. It
introduces a little code duplication, but the alternative is so horrible
that I really don't think it's worth it.