Build docs in CI
We recently had a failed release of the DB extractor due to a mistake in
the schema definition that wasn't caught before we tried to build the
docs during the final steps of the release process, leading to a
half-released version that it was very annoying to clean up.

Building the docs as part of CI would have caught that error.