Refactor string converter (#832)
This type is the result of years of feature creep, and it is incredibly messy. This splits the ConvertToString method into ConvertToJson and ConvertToString, which hopefully will make it a lot more readable. It introduces a little code duplication, but the alternative is so horrible that I really don't think it's worth it.

I originally wrote this code years ago, and it was largely based on trial and error. I now have a much better understanding of OPC-UA encoding, and the implementation is largely based on that.

In OPC-UA, you have the `Variant` type, which is used as the general "Value" type for any place where the type is dynamic. It can be a bunch of primitive types, i.e. string, short, long, bool, float, etc. but also a few structured types like `LocalizedText`, `QualifiedName`, `DiagnosticInfo`. See https://reference.opcfoundation.org/Core/Part6/v105/docs/5.2.2.16

In addition to this fixed list of builtin types, it can also be `ExtensionObject` which is an even more dynamic wrapper around thousands of different kinds of structures. Some of these will be deserialized by the SDK already, so we can convert them specifically to JSON.

The new implementation reflects this, and explicitly handles the valid variant values, with extra handling for a few ExtensionObject types.