Do not add state store restoration as an issue for new states (#892)
* Do not add state store restoration as an issue for new states

If there are new states, we will trigger history based on that later.
We were adding and removing a history issue during state restoration.
This caused history to restart after rebrowse, even when unnecessary.
Instead, I'll make two changes:

 - Only add new states to the history reader _after_ restoring state, if
   state restoration is enabled.
 - Add the state restoration history issue only on startup, so that we
   don't run history until we've done the _initial_ state restoration.

This probably needs a rethink, but this is all changing a lot with the
v3 refactor, and I'd prefer to wait with the big refactoring changes
until then. This should fix the issue.

* Try to improve test

* Bump version