Remove support for influxdb (#903)
This completely removes support for influxdb from the extractor. This was merged into v3-dev in #631.

This feature is not used, and reflects a different era for cognite, where we built features less in line with product goals and more in line with specific customer demands.

This is a major reduction of complexity. The PR itself is really simple, since it's mostly just deletion.


* Gemini